+++ staging_service/import_specifications/individual_parsers.py
@@ -129,25 +129,11 @@ def _normalize_xsv(val: str) -> PRIMITIVE_TYPE:
         return val if val else None
 
 
-def _check_for_duplicate_headers(
-    headers: pandas.Index, spec_source: SpecificationSource, header_index=0
-):
-    seen = set()
-    for name in headers.get_level_values(header_index):
-        if name in seen:
-            raise _ParseException(Error(
-                ErrorType.PARSE_FAIL,
-                f"Duplicate column name: {name.split('.')[0]}",
-                spec_source
-            ))
-        seen.add(name)
-
-
 def _normalize_headers(
-    headers: list[str], line_number: int, spec_source: SpecificationSource
+    headers: list[Any], line_number: int, spec_source: SpecificationSource
 ) -> list[str]:
     seen = set()
-    ret = [s.strip() for s in headers]
+    ret = [str(s).strip() if not pandas.isna(s) else None for s in headers]
     for i, name in enumerate(ret, start=1):
         if not name:
             raise _ParseException(Error(
@@ -164,21 +150,6 @@ def _normalize_headers(
             ))
         seen.add(name)
     return ret
-
-
-def _process_dataframe(df: pandas.DataFrame, spec_source: SpecificationSource, header_index=0
-) -> ParseResult:
-    results = []
-    recs: list[dict[tuple[str, ...], PRIMITIVE_TYPE]] = df.to_dict(orient="records")
-    for r in recs:
-        results.append(frozendict(
-            # headers is a tuple of the column headers
-            {headers[header_index].strip(): _normalize_pandas(val) for headers, val in r.items()}
-        ))
-    if not results:
-        raise _ParseException(Error(
-            ErrorType.PARSE_FAIL, "No non-header data in file", spec_source))
-    return ParseResult(spec_source, tuple(results))
 
 
 def _parse_xsv(path: Path, sep: str) -> ParseResults:
@@ -228,39 +199,41 @@ def parse_tsv(path: Path) -> ParseResults:
     return _parse_xsv(path, "\t")
 
 
-def _load_excel_tab_to_dataframe(excel: pandas.ExcelFile, spcsrc: SpecificationSource
-) -> pandas.DataFrame:
-    try:
-        return excel.parse(sheet_name=spcsrc.tab, header=[0, 1, 2])
-    except IndexError:
-        # ugh. https://github.com/pandas-dev/pandas/issues/43143
-        # I really hope I'm not swallowing other pandas bugs here, but not really
-        # any way to tell
-        # Not fixed as of pandas 1.4.0
-        raise _ParseException(Error(
-            ErrorType.PARSE_FAIL, "Missing expected header rows", spcsrc
-        ))
-
+def _process_excel_row(
+    row: tuple[Any], rownum: int, expected_columns: int, spcsrc: SpecificationSource
+) -> list[Any]:
+    while len(row) > expected_columns:
+        if pandas.isna(row[-1]):  # inefficient, but premature optimization...
+            row = row[0:-1]
+        else:
+            raise _ParseException(Error(
+                ErrorType.INCORRECT_COLUMN_COUNT,
+                f"Incorrect number of items in line {rownum}, "
+                + f"expected {expected_columns}, got {len(row)}",
+                spcsrc))
+    return row
+            
 
 def _process_excel_tab(excel: pandas.ExcelFile, spcsrc: SpecificationSource
 ) -> (O[str], O[ParseResult]):
-    df = _load_excel_tab_to_dataframe(excel, spcsrc)
-    if df.empty:  # might as well not error check headers in sheets with no data
-        return (None, None)  # next line will throw an error for a completely empty sheet
-    # at this point we know that the 3 headers are present
+    df = excel.parse(sheet_name=spcsrc.tab)
+    if df.shape[0] < 3:  # might as well not error check headers in sheets with no data
+        return (None, None)
+    # at this point we know that at least 4 lines are present - expecting the data type header,
+    # parameter ID header, display name header, and at least one data row
     header = df.columns.get_level_values(0)[0]
     datatype, columns = _parse_header(header, spcsrc, _VERSION)
-    # Could run through the entire sheet looking for non-NaN values to find offending rows
-    # Seems like a lot of complexity. If this becomes a problem reconsider
-    # Can't catch some header errors anyway, since pandas duplicates them silently
-    if df.columns.shape[0] != columns:
-        raise _ParseException(Error(
-            ErrorType.INCORRECT_COLUMN_COUNT,
-            f"Expected {columns} data columns, got {df.columns.shape[0]}",
-            spcsrc,
-        ))
-    _check_for_duplicate_headers(df.columns, spcsrc, header_index=1)
-    return datatype, _process_dataframe(df, spcsrc, header_index=1)
+    it = df.itertuples(index=False, name=None)
+    hd1 = _process_excel_row(next(it), 2, columns, spcsrc)
+    param_ids = _normalize_headers(hd1, 2, spcsrc)
+    _process_excel_row(next(it), 3, columns, spcsrc)
+    results = []
+    for i, row in enumerate(it, start=4):
+        row = _process_excel_row(row, i, columns, spcsrc)
+        if any(map(lambda x: not pandas.isna(x), row)):  # skip empty rows
+            results.append(frozendict(
+                {param_ids[j]: _normalize_pandas(row[j]) for j in range(len(row))}))
+    return datatype, ParseResult(spcsrc, tuple(results))
 
 
 def parse_excel(path: Path) -> ParseResults: